VAKIL MOSQUE
Masjed-e Vakil) is a mosque in Shiraz, in south of Iran, situated to the west of the Vakil Bazaar next to its entrance. This mosque was built between 1751 and 1773, during the Zand period; however, it was restored in the 19th century during the Qajar period. Shiraz was the seat of Karim Khan’s government and he endowed many buildings, including this mosque. 
Vakil Mosque covers an area of 8,660 square meters. It has only two iwans instead of the usual four, on the northern and southern sides of a large open court. The iwans and court are decorated with typical Shirazi seven colors tiles, a characteristic feature of the art and industry of Shiraz during the latter half of the 18th century. Its prayer hall, with an area of approximately 2,700 square meters, contains 48 monolithic pillars carved in spirals. The minbar in this hall is cut from a solid piece of green marble with a flight of 14 steps and is considered to be one of the master pieces of the Zand period. The exuberant floral decorative tiles largely date from the Qajar period.